Split tests for validation and rename rubric
